SDK-1300: Move "service" items into "docscan", to save passing same params around, and to protect us in case we want to make future breaking changes between these two packages
