+++ docscan/client.go
@@ -2,16 +2,19 @@ package docscan
 
 import (
 	"crypto/rsa"
+	"encoding/json"
+	"errors"
+	"io/ioutil"
 	"net/http"
 	"os"
 
 	"github.com/getyoti/yoti-go-sdk/v3/cryptoutil"
-	"github.com/getyoti/yoti-go-sdk/v3/docscan/service"
 	"github.com/getyoti/yoti-go-sdk/v3/docscan/session/create"
 	"github.com/getyoti/yoti-go-sdk/v3/docscan/session/retrieve"
 	"github.com/getyoti/yoti-go-sdk/v3/docscan/supported"
 	"github.com/getyoti/yoti-go-sdk/v3/media"
 	"github.com/getyoti/yoti-go-sdk/v3/requests"
+	"github.com/getyoti/yoti-go-sdk/v3/yotierror"
 )
 
 // Client is responsible for setting up test data in the sandbox instance.
@@ -24,6 +27,8 @@ type Client struct {
 	HTTPClient requests.HttpClient
 	// API URL to use. This is not required, and a default will be set if not provided.
 	apiURL string
+	// Mockable JSON marshaler
+	jsonMarshaler jsonMarshaler
 }
 
 // NewClient constructs a Client object
@@ -35,60 +40,210 @@ func NewClient(sdkID string, key []byte) (*Client, error) {
 	}
 
 	return &Client{
-		SdkID: sdkID,
-		Key:   decodedKey,
+		SdkID:      sdkID,
+		Key:        decodedKey,
+		HTTPClient: http.DefaultClient,
+		apiURL:     getAPIURL(),
 	}, err
 }
 
 // OverrideAPIURL overrides the default API URL for this Yoti Client
-func (client *Client) OverrideAPIURL(apiURL string) {
-	client.apiURL = apiURL
+func (c *Client) OverrideAPIURL(apiURL string) {
+	c.apiURL = apiURL
 }
 
-func (client *Client) getAPIURL() string {
-	if client.apiURL == "" {
-		if value, exists := os.LookupEnv("YOTI_DOC_SCAN_API_URL"); exists && value != "" {
-			client.apiURL = value
-		} else {
-			client.apiURL = "https://api.yoti.com/sandbox/idverify/v1"
-		}
+func getAPIURL() string {
+	if value, exists := os.LookupEnv("YOTI_DOC_SCAN_API_URL"); exists && value != "" {
+		return value
+	} else {
+		return "https://api.yoti.com/sandbox/idverify/v1"
 	}
-	return client.apiURL
-}
-
-func (client *Client) getHTTPClient() requests.HttpClient {
-	if client.HTTPClient != nil {
-		return client.HTTPClient
-	}
-	return http.DefaultClient
 }
 
 // CreateSession creates a Doc Scan session using the supplied session specification
-func (client *Client) CreateSession(sessionSpec *create.SessionSpecification) (*create.SessionResult, error) {
-	return service.CreateSession(client.getHTTPClient(), client.SdkID, client.Key, client.getAPIURL(), sessionSpec, nil)
+func (c *Client) CreateSession(sessionSpec *create.SessionSpecification) (*create.SessionResult, error) {
+	requestBody, err := marshalJSON(c.jsonMarshaler, sessionSpec)
+	if err != nil {
+		return nil, err
+	}
+
+	var request *http.Request
+	request, err = (&requests.SignedRequest{
+		Key:        c.Key,
+		HTTPMethod: http.MethodPost,
+		BaseURL:    c.apiURL,
+		Endpoint:   createSessionPath(),
+		Headers:    requests.JSONHeaders(),
+		Body:       requestBody,
+		Params:     map[string]string{"sdkID": c.SdkID},
+	}).Request()
+	if err != nil {
+		return nil, err
+	}
+
+	var response *http.Response
+	response, err = requests.Execute(c.HTTPClient, request, yotierror.DefaultHTTPErrorMessages)
+	if err != nil {
+		return nil, err
+	}
+
+	var responseBytes []byte
+	responseBytes, err = ioutil.ReadAll(response.Body)
+	if err != nil {
+		return nil, err
+	}
+
+	var result create.SessionResult
+	err = json.Unmarshal(responseBytes, &result)
+
+	return &result, err
 }
 
 // GetSession retrieves the state of a previously created Yoti Doc Scan session
-func (client *Client) GetSession(sessionID string) (*retrieve.GetSessionResult, error) {
-	return service.GetSession(client.getHTTPClient(), client.SdkID, client.Key, client.getAPIURL(), sessionID)
+func (c *Client) GetSession(sessionID string) (*retrieve.GetSessionResult, error) {
+	request, err := (&requests.SignedRequest{
+		Key:        c.Key,
+		HTTPMethod: http.MethodGet,
+		BaseURL:    c.apiURL,
+		Endpoint:   getSessionPath(sessionID),
+		Params:     map[string]string{"sdkID": c.SdkID},
+	}).Request()
+	if err != nil {
+		return nil, err
+	}
+
+	var response *http.Response
+	response, err = requests.Execute(c.HTTPClient, request, yotierror.DefaultHTTPErrorMessages)
+	if err != nil {
+		return nil, err
+	}
+
+	var responseBytes []byte
+	responseBytes, err = ioutil.ReadAll(response.Body)
+	if err != nil {
+		return nil, err
+	}
+
+	var result retrieve.GetSessionResult
+	err = json.Unmarshal(responseBytes, &result)
+
+	return &result, err
 }
 
 // DeleteSession deletes a previously created Yoti Doc Scan session and all of its related resources
-func (client *Client) DeleteSession(sessionID string) error {
-	return service.DeleteSession(client.getHTTPClient(), client.SdkID, client.Key, client.getAPIURL(), sessionID)
+func (c *Client) DeleteSession(sessionID string) error {
+	request, err := (&requests.SignedRequest{
+		Key:        c.Key,
+		HTTPMethod: http.MethodDelete,
+		BaseURL:    c.apiURL,
+		Endpoint:   deleteSessionPath(sessionID),
+		Params:     map[string]string{"sdkID": c.SdkID},
+	}).Request()
+	if err != nil {
+		return err
+	}
+
+	_, err = requests.Execute(c.HTTPClient, request, yotierror.DefaultHTTPErrorMessages)
+	if err != nil {
+		return err
+	}
+
+	return nil
 }
 
 // GetMediaContent retrieves media related to a Yoti Doc Scan session based on the supplied media ID
-func (client *Client) GetMediaContent(sessionID, mediaID string) (*media.Media, error) { // TODO: change to media.Value
-	return service.GetMediaContent(client.getHTTPClient(), client.SdkID, client.Key, client.getAPIURL(), sessionID, mediaID)
+func (c *Client) GetMediaContent(sessionID, mediaID string) (*media.Media, error) { // TODO: change to media.Value
+	request, err := (&requests.SignedRequest{
+		Key:        c.Key,
+		HTTPMethod: http.MethodGet,
+		BaseURL:    c.apiURL,
+		Endpoint:   getMediaContentPath(sessionID, mediaID),
+		Params:     map[string]string{"sdkID": c.SdkID},
+	}).Request()
+	if err != nil {
+		return nil, err
+	}
+
+	var response *http.Response
+	response, err = requests.Execute(c.HTTPClient, request, yotierror.DefaultHTTPErrorMessages)
+	if err != nil {
+		return nil, err
+	}
+
+	var responseBytes []byte
+	responseBytes, err = ioutil.ReadAll(response.Body)
+	if err != nil {
+		return nil, err
+	}
+
+	contentType := response.Header.Get("Content-type") // TODO: check this
+	if contentType == "" {
+		err = errors.New("unable to parse content type from response")
+	}
+
+	media := media.NewMedia(contentType, responseBytes)
+	return &media, err
 }
 
 // DeleteMediaContent deletes media related to a Yoti Doc Scan session based on the supplied media ID
-func (client *Client) DeleteMediaContent(sessionID, mediaID string) error {
-	return service.DeleteMediaContent(client.getHTTPClient(), client.SdkID, client.Key, client.getAPIURL(), sessionID, mediaID)
+func (c *Client) DeleteMediaContent(sessionID, mediaID string) error {
+	request, err := (&requests.SignedRequest{
+		Key:        c.Key,
+		HTTPMethod: http.MethodDelete,
+		BaseURL:    c.apiURL,
+		Endpoint:   deleteMediaPath(sessionID, mediaID),
+		Params:     map[string]string{"sdkID": c.SdkID},
+	}).Request()
+	if err != nil {
+		return err
+	}
+
+	_, err = requests.Execute(c.HTTPClient, request, yotierror.DefaultHTTPErrorMessages)
+	if err != nil {
+		return err
+	}
+
+	return nil
 }
 
 // GetSupportedDocuments gets a list of supported documents
-func (client *Client) GetSupportedDocuments() (*supported.DocumentsResponse, error) {
-	return service.GetSupportedDocuments(client.getHTTPClient(), client.Key, client.getAPIURL())
+func (c *Client) GetSupportedDocuments() (*supported.DocumentsResponse, error) {
+	request, err := (&requests.SignedRequest{
+		Key:        c.Key,
+		HTTPMethod: http.MethodGet,
+		BaseURL:    c.apiURL,
+		Endpoint:   getSupportedDocumentsPath(),
+	}).Request()
+	if err != nil {
+		return nil, err
+	}
+
+	var response *http.Response
+	response, err = requests.Execute(c.HTTPClient, request, yotierror.DefaultHTTPErrorMessages)
+	if err != nil {
+		return nil, err
+	}
+
+	var responseBytes []byte
+	responseBytes, err = ioutil.ReadAll(response.Body)
+	if err != nil {
+		return nil, err
+	}
+
+	var result supported.DocumentsResponse
+	err = json.Unmarshal(responseBytes, &result)
+
+	return &result, err
+}
+
+// jsonMarshaler is a mockable JSON marshaler
+type jsonMarshaler interface {
+	Marshal(v interface{}) ([]byte, error)
+}
+
+func marshalJSON(jsonMarshaler jsonMarshaler, v interface{}) ([]byte, error) {
+	if jsonMarshaler != nil {
+		return jsonMarshaler.Marshal(v)
+	}
+	return json.Marshal(v)
 }
